feat(interface-types) Move `serde::InterfaceTypeIterator` into `values::FlattenInterfaceValueIterator`.
